# core/openapi.yaml
openapi: 3.0.3
info:
  title: Pen CMS API Specification
  description: |
    The contract-first API specification for Pen CMS, a Git-native flat-file CMS.

    Humans and agents share one capability vocabulary (string lists). `write:*` is
    create+update (never split create:/update:). `delete:*` is separate. `read` is
    monolithic. `publish:content` is content status (PATCH approve/publish) and is
    not host `publish` (deploy + Deploy Grant). Generic PUT of a post/page does not
    sniff status — that stays under write:posts / write:pages.

    Legacy agent scope `write` expands one-way to every write:*, every delete:*, and
    publish:content. It never implies host publish, users:manage, or manage:sites.

    Human admin routes under /api authorize from cookie `pen_jwt` or Bearer JWT →
    user YAML. Forgeable `pen_role` / `pen_user_id` cookies are not a root of trust.
  version: 0.1.0
servers:
  - url: /api/v1
paths:
...
  /translations/config:
    get:
      summary: Get Translation Configuration
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/PenSiteId'
      responses:
        '200':
          description: Active site's translation configuration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslationConfig'
    put:
      summary: Update Translation Configuration
      description: Human/operator update of the existing site i18n registry fields.
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/PenSiteId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TranslationConfigUpdate'
      responses:
        '200':
          description: Updated translation configuration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslationConfig'
        '400':
          description: Invalid language configuration.
        '403':
          description: Agent tokens cannot alter site translation configuration.
  /translations/coverage:
    get:
      summary: Get Exact Translation Coverage
      description: >-
        Count exact sibling files only. Merged-list fallback rows never count as
        coverage.
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/PenSiteId'
        - name: language
          in: query
          required: false
          schema:
            type: string
          description: Optional exact target-language filter.
      responses:
        '200':
          description: Exact coverage totals and item gaps for the active site.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslationCoverageResponse'
        '400':
          description: Invalid or unconfigured language.
  /translations/strings:
    get:
      summary: Get Effective UI Strings
      description: >-
        Resolve file-backed reader chrome strings for one exact configured
        language using engine, theme, site-default, and site-target layers.
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/PenSiteId'
        - name: language
          in: query
          required: false
          schema:
            type: string
          description: Exact configured language. Omitted means the site default.
      responses:
        '200':
          description: Effective strings, source badges, and sparse site overrides.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UiStringBundle'
        '400':
          description: Invalid or unconfigured language.
    put:
      summary: Replace Sparse UI String Overrides
      description: >-
        Human/operator replacement of one language's sparse site override file.
        Omitting a key resets it to the next fallback layer.
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/PenSiteId'
        - name: language
          in: query
          required: true
          schema:
            type: string
          description: Exact configured language whose site override file is replaced.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UiStringOverrides'
      responses:
        '200':
          description: Updated effective strings and sparse overrides.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UiStringBundle'
        '400':
          description: Inactive i18n, invalid language, key, or non-string value.
        '403':
          description: Human admin required.
  /translations/{slug}/{language}/review:
    post:
      summary: Review Translation Sibling
      description: >-
        Approve promotes draft/stub to published and clears review. Reject
        withholds the sibling and records the decision without deleting it.
        Agent tokens cannot self-review; Autonomous agents set status on write.
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/PenSiteId'
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: language
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TranslationReviewRequest'
      responses:
        '200':
          description: Review transition recorded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntryMutationResponse'
        '403':
          description: Human review is required by the current publication policy.
        '404':
          description: Exact sibling not found.
        '409':
          description: Sibling is not in a reviewable state.
  /translations/runs:
    get:
      summary: List Translation Runs
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/PenSiteId'
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
      responses:
        '200':
          description: Bounded recent external run telemetry.
          content:
            application/json:
              schema:
                type: object
                required: [runs]
                properties:
                  runs:
                    type: array
                    items:
                      $ref: '#/components/schemas/TranslationRun'
    post:
      summary: Start Translation Run
      description: Start telemetry for an external cron or MCP run; this does not schedule work.
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/PenSiteId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TranslationRunStart'
      responses:
        '201':
          description: Run telemetry started.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslationRun'
        '403':
          description: Site-bound authorization or agent pause denied the run.
  /translations/runs/{run_id}:
    patch:
      summary: Update or Finish Translation Run
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/PenSiteId'
        - name: run_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TranslationRunUpdate'
      responses:
        '200':
          description: Run telemetry updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslationRun'
        '403':
          description: Caller does not own this site-bound run.
        '404':
          description: Run not found.
  /mcp/translations/config:
    get:
      summary: MCP Get Translation Configuration
      description: HTTP surface behind the get_translation_config MCP tool.
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Site-bound translation configuration and policy health.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslationConfig'
...
  /mcp/translations/{slug}/{language}/review:
    post:
      summary: MCP Review Translation Sibling
      description: Human-only review parity surface; agent tokens cannot self-review.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: language
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              additionalProperties: false
              properties:
                decision:
                  type: string
                  enum: [approve, reject]
      responses:
        '200':
          description: Review transition recorded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntryMutationResponse'
        '403':
          description: Human review is required.
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Pen-API-Key
    CookieJwtAuth:
      type: apiKey
      in: cookie
      name: pen_jwt
      description: Human session JWT. Authorize from this cookie → YAML. Never trust pen_role / pen_user_id.
    BearerJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Human or agent JWT. Agent tokens use type=agent + scopes + site_id; they do not inherit the sponsor admin role.
  parameters:
    PenSiteId:
      name: X-Pen-Site-Id
      in: header
      required: false
      schema:
        type: string
        default: default
      description: >
        Active Content site id for human requests. Honored only if the actor
        has membership on that site (or YAML role admin). Unknown ids → 400.
        Non-member → 403 site_access_denied. Agent JWTs ignore this header and
        use the token site_id only.
...
  schemas:
...
    LanguageLabels:
      type: object
      default: {}
      additionalProperties:
        type: string
        minLength: 1
      description: Display-label overrides keyed by normalized lowercase BCP-47 tag.
    SiteI18nConfig:
      type: object
      properties:
        language:
          type: string
          default: en
          example: en
          description: Default language as a known, normalized lowercase BCP-47 tag.
        languages:
          type: array
          default: []
          uniqueItems: true
          description: >-
            Ordered normalized BCP-47 tags. When non-empty, this list must include
            language. i18n remains inactive until it contains at least two tags.
          items:
            type: string
            example: fr
        language_labels:
          $ref: '#/components/schemas/LanguageLabels'
        translation_automation_paused:
          type: boolean
          default: false
          description: Whether automated translation work is paused for this site.
...
    EntryDetail:
      type: object
      required:
        - frontmatter
        - body
      properties:
        frontmatter:
          type: object
          description: Key-value metadata matching collection specifications.
        body:
          type: string
          description: Markdown body content.
        composite:
          type: boolean
          nullable: true
          description: Omit on update to preserve the existing composite structure.
        partials:
          type: object
          nullable: true
          additionalProperties:
            type: string
        run_id:
          type: string
          nullable: true
          writeOnly: true
          description: >-
            Active bound telemetry run for an agent exact-sibling update.
            Required when automation_policy is enabled; ignored for reads.
        language:
          type: string
          description: Exact content language; present only when site i18n is active.
        translation_group:
          type: string
          nullable: true
          pattern: '^tg_[0-9a-f]{32}$'
        translations:
          type: array
          items:
            $ref: '#/components/schemas/TranslationPeer'
        provenance:
          $ref: '#/components/schemas/ContentProvenance'
        version:
          type: string
          nullable: true
          readOnly: true
          description: >-
            Opaque optimistic-concurrency token from the content file mtime.
            Echo it back as expected_version on a later PUT.
        expected_version:
          type: string
          nullable: true
          writeOnly: true
          description: >-
            Opaque token from a prior GET `version`. A mismatch returns 409
            version_conflict unless force is true. Omitted means unconditional write.
        force:
          type: boolean
          writeOnly: true
          default: false
          description: Explicit overwrite after a version_conflict. Omitted expected_version is not force.
    TranslationPeer:
      type: object
      required:
        - language
        - published
        - needs_review
      properties:
        language:
          type: string
        status:
          type: string
          nullable: true
        published:
          type: boolean
        needs_review:
          type: boolean
        review_decision:
          type: string
          nullable: true
          enum: [pending, approved, rejected]
    ContentProvenance:
      type: object
      readOnly: true
      required:
        - created_by
        - needs_review
      properties:
        created_by:
          type: string
          enum: [human, agent]
        created_by_id:
          type: string
          nullable: true
        updated_by:
          type: string
          nullable: true
          enum: [human, agent]
        updated_by_id:
          type: string
          nullable: true
        run_id:
          type: string
          nullable: true
        needs_review:
          type: boolean
        reviewed_by:
          type: string
          nullable: true
        reviewed_at:
          type: string
          format: date-time
          nullable: true
        review_decision:
          type: string
          enum: [pending, approved, rejected]
        review_note:
          type: string
          nullable: true
          maxLength: 500
    EntryMutationResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
        version:
          type: string
          nullable: true
          description: Opaque mtime token of the file after a successful write.
        version_warning:
          type: string
          description: Present only when expected_version mismatched and strict 409 is disabled.
        entry:
          description: Present for active-i18n exact-language mutations; omitted by legacy monolingual writes.
          allOf:
            - $ref: '#/components/schemas/EntryDetail'
            - type: object
              properties:
                provenance:
                  $ref: '#/components/schemas/ContentProvenance'
...
    TranslationReviewRequest:
      type: object
      required:
        - decision
      additionalProperties: false
      properties:
        decision:
          type: string
          enum: [approve, reject]
        note:
          type: string
          maxLength: 500
    TranslationConfig:
      allOf:
        - $ref: '#/components/schemas/SiteI18nConfig'
        - type: object
          required:
            - language
            - languages
            - language_labels
            - translation_automation_paused
            - i18n_active
            - automation_policy
          properties:
            i18n_active:
              type: boolean
              readOnly: true
            automation_policy:
              $ref: '#/components/schemas/LocalizationAutomationPolicy'
    TranslationConfigUpdate:
      allOf:
        - $ref: '#/components/schemas/SiteI18nConfig'
        - type: object
          required:
            - language
            - languages
            - language_labels
            - translation_automation_paused
          properties:
            automation_policy:
              $ref: '#/components/schemas/LocalizationAutomationPolicy'
    LocalizationTargetPolicy:
      type: object
      required:
        - operation
        - model
        - agent_key_id
        - review_policy
      additionalProperties: false
      properties:
        operation:
          type: string
          enum: [translate, transliterate, translate_then_transliterate]
        model:
          type: string
          minLength: 1
          maxLength: 200
          description: Non-secret model identifier interpreted by the external caller.
        agent_key_id:
          type: string
          description: Immutable non-secret ID of the site-bound agent key (granular scopes or legacy read/write/publish).
        review_policy:
          type: string
          enum: [require_review, allow_unreviewed_draft]
          default: require_review
        agent_key_name:
          type: string
          nullable: true
          readOnly: true
        binding_valid:
          type: boolean
          readOnly: true
        binding_error:
          type: string
          nullable: true
          readOnly: true
          enum: [missing_or_revoked, wrong_site, missing_scope]
    LocalizationAutomationPolicy:
      type: object
      required:
        - enabled
        - targets
      additionalProperties: false
      properties:
        enabled:
          type: boolean
          default: false
        targets:
          type: object
          description: Policies keyed by configured non-default language.
          additionalProperties:
            $ref: '#/components/schemas/LocalizationTargetPolicy'
        policy_valid:
          type: boolean
          readOnly: true
        policy_error:
          type: string
          nullable: true
          readOnly: true
    TranslationCoverageTotals:
      type: object
      required:
        - eligible
        - existing
        - published
        - draft
        - needs_review
        - rejected
        - missing
      properties:
        eligible:
          type: integer
          minimum: 0
        existing:
          type: integer
          minimum: 0
        published:
          type: integer
          minimum: 0
        draft:
          type: integer
          minimum: 0
        needs_review:
          type: integer
          minimum: 0
        rejected:
          type: integer
          minimum: 0
        missing:
          type: integer
          minimum: 0
    TranslationCoverageSibling:
      allOf:
        - $ref: '#/components/schemas/TranslationPeer'
        - type: object
          required:
            - translation_group
          properties:
            translation_group:
              type: string
              nullable: true
    TranslationCoverageItem:
      type: object
      required:
        - slug
        - collection
        - source
        - siblings
        - gap_codes
      properties:
        slug:
          type: string
        collection:
          type: string
          description: Exact collection required by content sibling routes.
        source:
          $ref: '#/components/schemas/TranslationCoverageSibling'
        siblings:
          type: array
          items:
            $ref: '#/components/schemas/TranslationCoverageSibling'
        gap_codes:
          type: array
          items:
            type: string
    TranslationCoverageResponse:
      type: object
      required:
        - config
        - totals
        - items
      properties:
        config:
          $ref: '#/components/schemas/TranslationConfig'
        totals:
          $ref: '#/components/schemas/TranslationCoverageTotals'
        items:
          type: array
          items:
            $ref: '#/components/schemas/TranslationCoverageItem'
    UiStringValue:
      type: object
      required:
        - effective
        - source
      properties:
        effective:
          type: string
        source:
          type: string
          enum: [engine, theme, site_default, site_target]
        override:
          type: string
          nullable: true
          description: Sparse value stored in the requested language's site file.
    UiStringBundle:
      type: object
      required:
        - config
        - language
        - strings
        - overrides
      properties:
        config:
          $ref: '#/components/schemas/TranslationConfig'
        language:
          type: string
        strings:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/UiStringValue'
        overrides:
          type: object
          additionalProperties:
            type: string
    UiStringOverrides:
      type: object
      required:
        - overrides
      additionalProperties: false
      properties:
        overrides:
          type: object
          additionalProperties:
            type: string
    TranslationRunCounts:
      type: object
      additionalProperties: false
      properties:
        discovered:
          type: integer
          minimum: 0
        attempted:
          type: integer
          minimum: 0
        created:
          type: integer
          minimum: 0
        updated:
          type: integer
          minimum: 0
        skipped:
          type: integer
          minimum: 0
        failed:
          type: integer
          minimum: 0
    TranslationRunStart:
      type: object
      required:
        - mode
      additionalProperties: false
      properties:
        mode:
          type: string
          enum: [translate, transliterate, translate_then_transliterate, manual]
        target_languages:
          type: array
          uniqueItems: true
          maxItems: 32
          items:
            type: string
    TranslationRunUpdate:
      type: object
      additionalProperties: false
      properties:
        status:
          type: string
          enum: [running, completed, failed, cancelled]
        counts:
          $ref: '#/components/schemas/TranslationRunCounts'
        error:
          type: string
          maxLength: 160
    TranslationRun:
      type: object
      required:
        - run_id
        - site_id
        - actor
        - actor_id
        - mode
        - target_languages
        - policy_applied
        - review_policy
        - started_at
        - status
        - counts
      properties:
        run_id:
          type: string
        site_id:
          type: string
        actor:
          type: string
          enum: [human, agent]
        actor_id:
          type: string
        mode:
          type: string
        target_languages:
          type: array
          items:
            type: string
        policy_applied:
          type: boolean
          description: Whether this run was validated against an enabled target policy.
        model:
          type: string
          nullable: true
          description: Non-secret model identifier snapshotted at run start.
        agent_key_id:
          type: string
          nullable: true
          description: Immutable non-secret key ID snapshotted at run start.
        agent_key_name:
          type: string
          nullable: true
        review_policy:
          type: string
          enum: [require_review, allow_unreviewed_draft]
        started_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        finished_at:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [running, completed, failed, cancelled]
        counts:
          $ref: '#/components/schemas/TranslationRunCounts'
        error:
          type: string
          nullable: true
...